update build config
