+++ quickstart/prepare-example.sh
@@ -72,7 +72,8 @@ EXTENSIONS="pdo"
 EXTENSIONS="${EXTENSIONS},pgsql,pdo_pgsql"
 
 ./bin/spc build:libs "libxml2"  --cc=clang --cxx=clang++ --debug
-
+exit 0
+exit 0
 ./bin/spc build:libs "postgresql"  --cc=clang --cxx=clang++ --debug
 exit 0
 ./bin/spc build:libs "libiconv,libxml2,zstd,zlib,openssl,ncurses,readline,icu,postgresql"  --cc=clang --cxx=clang++ --debug
