+++ src/SPC/builder/macos/library/libxml2.php
@@ -4,45 +4,11 @@ declare(strict_types=1);
 
 namespace SPC\builder\macos\library;
 
-use SPC\exception\RuntimeException;
-use SPC\store\FileSystem;
+use SPC\builder\unix\library\libxml2 as libxml;
 
 class libxml2 extends MacOSLibraryBase
 {
+    use libxml;
+
     public const NAME = 'libxml2';
-
-    /**
-     * @throws RuntimeException
-     */
-    protected function build()
-    {
-        $enable_zlib = $this->builder->getLib('zlib') ? 'ON' : 'OFF';
-        $enable_icu = $this->builder->getLib('icu') ? 'ON' : 'OFF';
-        $enable_xz = $this->builder->getLib('xz') ? 'ON' : 'OFF';
-
-        [$lib, $include, $destdir] = SEPARATED_PATH;
-
-        FileSystem::resetDir($this->source_dir . '/build');
-        shell()->cd($this->source_dir . '/build')
-            ->exec(
-                "{$this->builder->configure_env} " . ' cmake ' .
-                // '--debug-find ' .
-                '-DCMAKE_BUILD_TYPE=Release ' .
-                '-DBUILD_SHARED_LIBS=OFF ' .
-                '-DLIBXML2_WITH_ICONV=ON ' .
-                "-DLIBXML2_WITH_ZLIB={$enable_zlib} " .
-                "-DLIBXML2_WITH_ICU={$enable_icu} " .
-                "-DLIBXML2_WITH_LZMA={$enable_xz} " .
-                '-DLIBXML2_WITH_PYTHON=OFF ' .
-                '-DLIBXML2_WITH_PROGRAMS=OFF ' .
-                '-DLIBXML2_WITH_TESTS=OFF ' .
-                '-DCMAKE_INSTALL_PREFIX=/ ' .
-                "-DCMAKE_INSTALL_LIBDIR={$lib} " .
-                "-DCMAKE_INSTALL_INCLUDEDIR={$include} " .
-                "-DCMAKE_TOOLCHAIN_FILE={$this->builder->cmake_toolchain_file} " .
-                '..'
-            )
-            ->exec("cmake --build . -j {$this->builder->concurrency}")
-            ->exec("make install DESTDIR={$destdir}");
-    }
 }
